# swagger/index.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API documenations using Swagger",
    "description": "Sample project which generates API documentaion using swagger.",
    "contact": {
      "email": "princekfrancis@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
  "paths": {
    "/contact": {
      "get": {
        "tags": [
          "Contact"
        ],
        "summary": "Get all the contacts",
        "description": "Get all the contacts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Contacts fetched successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          },
          "400": {
            "description": "Error in fetching contacts",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contact"
        ],
        "summary": "Save a new contact",
        "description": "Save a new contact",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Contact object",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact saved successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          },
          "400": {
            "description": "Error in saving contact",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      }
    },
    "/contact/{id}": {
      "get": {
        "tags": [
          "Contact"
        ],
        "summary": "Get a contact",
        "description": "Get a contact with the id specified in parameter",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Contact id that needs to be fetched",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact fetched successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Contact"
            }
          },
          "400": {
            "description": "Error in fetching contact",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contact"
        ],
        "summary": "Update a contact",
        "description": "Update a contact",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Contact id that needs to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Contact object",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          },
          "400": {
            "description": "Error in updating contact",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contact"
        ],
        "summary": "Delete a contact",
        "description": "Delete a contact with the id specified in parameter",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Contact id that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          },
          "400": {
            "description": "Error in deleting contact",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      }
    },
    "/product": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get all the products",
        "description": "Get all the products",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Products fetched successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Error in fetching products",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Save a new product",
        "description": "Save a new product",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product saved successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          },
          "400": {
            "description": "Error in saving product",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get a product",
        "description": "Get a product with the id specified in parameter",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Product id that needs to be fetched",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product fetched successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Error in fetching product",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "Update a product",
        "description": "Update a product",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Product id that needs to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Product object",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          },
          "400": {
            "description": "Error in updating product",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Delete a product",
        "description": "Delete a product with the id specified in parameter",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Product id that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          },
          "400": {
            "description": "Error in deleting product",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CommonResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "Contact": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "mobile": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        }
      }
    }
  }
}